Restore workflow file as existed prior to PR #635 (#647)
### Motivation and Context

This restores a change required in PR #635 -- comment out the freeze on
the yarn.lock file during install yarn in the workflow. This is because
the file that needed to be changed in the PR was the yarn.lock file
itself.

### Description

This removes the # that was addin in PR #635 to the workflow file.

---------

Signed-off-by: dependabot[bot] <support@github.com>
Co-authored-by: dependabot[bot] <49699333+dependabot[bot]@users.noreply.github.com>
Co-authored-by: Mollie Munoz <momuno@microsoft.com>